pyproject.toml
[project]
name = "hashbidder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.2",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "D"]
ignore = ["D105"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
strict = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "hypothesis>=6.151.11",
    "mypy>=1.20.0",
    "pytest>=9.0.2",
    "ruff>=0.15.9",
]

[project.scripts]
hashbidder = "hashbidder.main:main"

